requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.54
botocore==1.35.54
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.0
eval_type_backport==0.2.0
fastapi==0.115.4
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.6
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.7
langchain-google-genai==2.0.0
langchain-openai==0.2.1
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.139
lxml==5.3.0
mangum==0.19.0
MarkupSafe==3.0.2
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.53.0
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pinecone==4.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyparsing==3.2.0
pypdf==5.1.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.66.6
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.4
unstructured-client==0.26.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.17.1
